Reviewed comments - rename Actor directory to MapObj and small changes
