+++ include/Game/al/Interfaces.h
@@ -27,11 +27,6 @@ namespace al {
 
     class IUseNerve {
     public:
-        inline IUseNerve()
-        {
-            
-        }
-
         virtual NerveKeeper* getNerveKeeper() const = 0;
     };
